// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  TEACHER: 'TEACHER',
  STUDENT: 'STUDENT'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Rarity = {
  COMMON: 'COMMON',
  RARE: 'RARE',
  EPIC: 'EPIC',
  LEGENDARY: 'LEGENDARY'
} as const

export type Rarity = (typeof Rarity)[keyof typeof Rarity]


export const Reason = {
  BEHAVIOR: 'BEHAVIOR',
  TASK: 'TASK',
  ATTENDANCE: 'ATTENDANCE',
  PARTICIPATION: 'PARTICIPATION',
  OTHER: 'OTHER'
} as const

export type Reason = (typeof Reason)[keyof typeof Reason]